Fix show more button bug in case studies page (#172)
This fixes the show more button functionality by ensuring a consistent
default number of visible tags (5)

Context
The show more button on the case studies page was not working correctly
because the default number of visible tags was not properly set in the
request data. This change ensures that the UI consistently shows 5 tags
by default before expanding.

Testing

- Verify that the show more button appears when there are more than 5
tags

- Verify that clicking the show more button expands to show all tags

- Verify that the number of initially visible tags is consistently 5